automated estimation stabilized
this is a copy from the stable version in
https://github.com/IALSA/ialsa-bivariate-growth/tree/master/sandbox/pipeline-demo-1

@ampiccinin , good news! The automatic estimation is stable. Mass
production of models is nigh. The next oder of business is to connect
the extraction and graphic facilities (which shouldn't take long).

@casslbrown , please peruse the scripts (`0-prepare-data.R` and
`1-run-models.R`) to get a feel for them.  If possible, try to
investigate why models have lower N than in the data prepared
previously. Let's go over the production line together on Tuesday.