movedd files to appropriate subfolder
@andreazammit , thank you for the upload! I have moved the models to a
separate folder, which would help us keep track of the files. in the
future, please enclose the submitted models into the subfolder in the
./studies/eas/ and please include the date of submission into the name
of the folder.

Also, I see that the models you have submitted today are the same as the
models you have submitted on 2016-11-11 (see
./studies/eas/physical-cognitive/without-errors/grip-2016-11-11). Is
that so? If not, could you please describe the nature of the difference
between these two submission, so that we can make an informative choice
of which ones to use?  Thanks again!