the new structure of cognitive domains
@eduggan @ampiccinin
Please make a note of this file. This is where all further encodings
regarding the structure of cogntivie domains should take place. If a
separate structure is required for a project, replace first componet of
the name: `gait-domain-structure-dead` or `grip-domain-strucutre-dead`.
Hopefully, this will not be necessary and this structure will be able to
carry through all three tracks.